+++ src/PSRule.Rules.Azure/rules/Azure.Cosmos.Rule.ps1
@@ -13,6 +13,45 @@ Rule 'Azure.Cosmos.DefenderCloud' -Ref 'AZR-000382' -Type 'Microsoft.DocumentDb/
     $Assert.GreaterOrEqual($defender, '.', 1).Reason($LocalizedData.SubResourceNotFound, 'Microsoft.Security/advancedThreatProtectionSettings')
 } -Configure @{ AZURE_COSMOS_DEFENDER_PER_ACCOUNT = $False }
 
+# Synopsis: Use zone redundant Cosmos DB accounts in supported regions to improve reliability.
+Rule 'Azure.Cosmos.AvailabilityZone' -Ref 'AZR-000502' -Type 'Microsoft.DocumentDb/databaseAccounts' -Tag @{ release = 'GA'; ruleSet = '2025_12'; 'Azure.WAF/pillar' = 'Reliability'; } -Labels @{ 'Azure.WAF/maturity' = 'L1' } {
+    # Check for availability zones based on virtual machine scale sets, because it is not exposed through the provider for Cosmos DB.
+    $provider = [PSRule.Rules.Azure.Runtime.Helper]::GetResourceType('Microsoft.Compute', 'virtualMachineScaleSets');
+
+    $Assert.GreaterOrEqual($TargetObject, 'properties.locations', 1);
+
+    if ($TargetObject.properties.locations) {
+        foreach ($location in $TargetObject.properties.locations) {
+            $availabilityZones = GetAvailabilityZone -Location $location.locationName -Zone $provider.ZoneMappings;
+
+            # If the location supports availability zones, ensure zone redundancy is enabled
+            if ($availabilityZones) {
+                $Assert.HasFieldValue($location, 'isZoneRedundant', $true).
+                ReasonFrom('properties.locations', $LocalizedData.CosmosDBAvailabilityZone, $TargetObject.Name, $location.locationName);
+            }
+        }
+    }
+}
+
+# Synopsis: Use zone redundant Cosmos DB vCore clusters in supported regions to improve reliability.
+Rule 'Azure.Cosmos.MongoAvailabilityZone' -Ref 'AZR-000503' -Type 'Microsoft.DocumentDB/mongoClusters' -Tag @{ release = 'GA'; ruleSet = '2025_12'; 'Azure.WAF/pillar' = 'Reliability'; } -Labels @{ 'Azure.WAF/maturity' = 'L1' } {
+    # Check for availability zones based on virtual machine scale sets, because it is not exposed through the provider for Cosmos DB.
+    $provider = [PSRule.Rules.Azure.Runtime.Helper]::GetResourceType('Microsoft.Compute', 'virtualMachineScaleSets');
+    
+    $location = $TargetObject.Location;
+    $availabilityZones = GetAvailabilityZone -Location $location -Zone $provider.ZoneMappings;
+
+    # If the location supports availability zones, ensure zone redundancy is enabled
+    if ($availabilityZones) {
+        $Assert.HasFieldValue($TargetObject, 'properties.highAvailability.targetMode', 'ZoneRedundantPreferred').
+        ReasonFrom('properties.highAvailability.targetMode', $LocalizedData.MongoDBvCoreAvailabilityZone, $TargetObject.Name, $location);
+    }
+    else {
+        # Pass if the region does not support availability zones
+        $Assert.Pass();
+    }
+}
+
 #endregion Rules
 
 #region Helper functions